requirements.txt
gradio
openai-whisper
translate